Keyboard: Implement an optional BootKeyboard fallback
Based on earlier work by @obra, this implements a `BootKeyboard` fallback. When
the boot keyboard is in boot report mode, dispatch pretty much everything to it.

With this setup, code using `Keyboard` will automatically support the
`BootKeyboard` fallback too.

We make it optional by having a `bootKeyboard` member defaulting to NULL. When
boot support is desirable, it should be set to `&BootKeyboard`, and the default
protocol to `HID_BOOT_PROTOCOL` if support for things like FreeBSD is needed.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>