Demo of boot protocol working
This adds serial debugging output showing when a boot protocol HID report is sent (and
when an NKRO protocol report is sent, too), and the contents of that report. It also
changes the variable used to decide which type of report to send to the host, without
changing the `getDescriptor()` and `setup()` methods, which still do whatever they do with
the `protocol` variable, and `UEDATX`.

I don't understand why this works, but it seems to allow proper switching between boot
protocol and report protocol on macOS & Linux (the two machines I have access to).