Improve handling of simultaneous changes in a single NKRO report
When `sendReport()` is called, it now first checks for changes in the modifiers
byte, and breaks down the process of transitioning to the new report into three
stages:

1. First, any non-modifier keycodes that toggled off are removed from the old
   report, if any, and a report is (conditionally) sent. This ensures that a key
   with a modifier flag attached that toggles off won't have the host process
   the modifier first, potentially causing a erroneous repeat character. For
   example, if `LSHIFT(Key_4)` is held and released, we need to first send a
   report without the `4`, or some hosts will process the release of `shift`
   first, and give us `$$$4` in the output instead of `$$$$`.

2. Second, the modifiers that toggled on and/or off are updated, and a report is
   sent. This handles the situation where a key with a modifier flag toggles on,
   and we need to ensure that the host will process the modifier first to
   prevent erroneous output on the first character (`4$$$` vs `$$$$`).

3. Last, if there are any non-modifier keycodes that toggled on, they are added
   to the old report, and the final report is sent.

To accomplish this, the old report is modified in each step, and
`sendReportUnchecked()` sends the contents of that (modified) old report rather
than the new one. We could avoid that by making a temporary copy of the old
report, and updating and sending that instead, but there isn't much point to
doing so because there isn't any reasonable way for us to recover from
errors. Kaleidoscope doesn't even check the return value when calling
`sendReport()`, and it's not clear what we should do if an error occurs,
anyway.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>