Change internal names of NKRO HID report variables
- `report_`      => `pending_report_`
- `last_report_` => `current_report_`

These terms better reflect how they're used internally. The "current" report is
now updated just before a report is sent, rather than after, and it's still
"current" until a subsequent report is sent. The "pending" report, on the other
hand, is often incomplete, so it would be best to give it a name that reflects
it's volatile nature.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>